refactor: Remove SourceService, move it to the GitModelService
