+++ frontend/src/app/projects/models/create-coworking-method/create-coworking-method.component.ts
@@ -41,11 +41,10 @@ import {
 } from 'src/app/services/settings/git-settings.service';
 import {
   CreateGitModel,
-  GitModel,
+  GetGitModel,
+  GitModelService,
   PatchGitModel,
-  SourceService,
-} from 'src/app/services/source/source.service';
-import { GitModelService } from '../../project-detail/model-overview/model-detail/git-model.service';
+} from '../../project-detail/model-overview/model-detail/git-model.service';
 
 @Component({
   selector: 'app-create-coworking-method',
@@ -85,7 +84,7 @@ export class CreateCoworkingMethodComponent implements OnInit, OnDestroy {
   });
 
   private gitModelId?: number;
-  public gitModel?: GitModel;
+  public gitModel?: GetGitModel;
 
   public isEditMode: boolean = false;
   public editing: boolean = false;
@@ -102,7 +101,6 @@ export class CreateCoworkingMethodComponent implements OnInit, OnDestroy {
     private gitSettingsService: GitSettingsService,
     private gitService: GitService,
     private gitModelService: GitModelService,
-    private sourceService: SourceService,
     private router: Router,
     private route: ActivatedRoute
   ) {}
@@ -241,7 +239,7 @@ export class CreateCoworkingMethodComponent implements OnInit, OnDestroy {
 
   onCreateSubmit(): void {
     if (this.form.valid) {
-      this.sourceService
+      this.gitModelService
         .addGitSource(
           this.projectService.project?.slug!,
           this.modelService.model?.slug!,
@@ -301,7 +299,7 @@ export class CreateCoworkingMethodComponent implements OnInit, OnDestroy {
     };
   }
 
-  private fillFormWithGitModel(gitModel: GitModel): void {
+  private fillFormWithGitModel(gitModel: GetGitModel): void {
     this.urls.inputUrl.setValue(gitModel.path);
 
     const credentials = this.form.controls.credentials.controls;
