+++ frontend/src/app/projects/project-detail/model-overview/model-detail/git-model.service.ts
@@ -6,10 +6,6 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { BehaviorSubject, Observable, Subject, tap } from 'rxjs';
-import {
-  GitModel,
-  PatchGitModel,
-} from 'src/app/services/source/source.service';
 import { environment } from 'src/environments/environment';
 
 @Injectable({
@@ -20,15 +16,15 @@ export class GitModelService {
 
   constructor(private http: HttpClient) {}
 
-  private _gitModel = new Subject<GitModel>();
-  private _gitModels = new BehaviorSubject<Array<GitModel>>([]);
+  private _gitModel = new Subject<GetGitModel>();
+  private _gitModels = new BehaviorSubject<Array<GetGitModel>>([]);
 
   readonly gitModel = this._gitModel.asObservable();
   readonly gitModels = this._gitModels.asObservable();
 
   loadGitModels(project_slug: string, model_slug: string): void {
     this.http
-      .get<Array<GitModel>>(
+      .get<Array<GetGitModel>>(
         this.BACKEND_URL_PREFIX +
           `/projects/${project_slug}/models/${model_slug}/git`
       )
@@ -41,7 +37,7 @@ export class GitModelService {
     git_model_id: number
   ): void {
     this.http
-      .get<GitModel>(
+      .get<GetGitModel>(
         this.BACKEND_URL_PREFIX +
           `/projects/${project_slug}/models/${model_slug}/git/${git_model_id}`
       )
@@ -53,9 +49,9 @@ export class GitModelService {
     model_slug: string,
     git_model_id: number,
     gitModel: PatchGitModel
-  ): Observable<GitModel> {
+  ): Observable<GetGitModel> {
     return this.http
-      .patch<GitModel>(
+      .patch<GetGitModel>(
         this.BACKEND_URL_PREFIX +
           `/projects/${project_slug}/models/${model_slug}/git/${git_model_id}`,
         gitModel
@@ -72,9 +68,9 @@ export class GitModelService {
     project_slug: string,
     model_slug: string,
     git_model_id: number
-  ): Observable<GitModel> {
+  ): Observable<GetGitModel> {
     return this.http
-      .patch<GitModel>(
+      .patch<GetGitModel>(
         this.BACKEND_URL_PREFIX +
           `/projects/${project_slug}/models/${model_slug}/git/${git_model_id}`,
         true
@@ -86,4 +82,37 @@ export class GitModelService {
         })
       );
   }
+
+  addGitSource(
+    project_slug: string,
+    model_slug: string,
+    source: CreateGitModel
+  ): Observable<GetGitModel> {
+    return this.http.post<GetGitModel>(
+      environment.backend_url +
+        `/projects/${project_slug}/models/${model_slug}/modelsources/git`,
+      source
+    );
+  }
+}
+
+export interface BaseGitModel {
+  path: string;
+  revision: string;
+  entrypoint: string;
+  username: string;
+}
+
+export interface CreateGitModel extends BaseGitModel {
+  password?: string;
+}
+
+export interface PatchGitModel extends CreateGitModel {
+  primary: boolean;
+}
+
+export interface GetGitModel extends BaseGitModel {
+  id: number;
+  primary: boolean;
+  password: boolean;
 }
