--- frontend/src/app/services/source/source.service.spec.ts
@@ -1,21 +1,0 @@
-/*
- * SPDX-FileCopyrightText: Copyright DB Netz AG and the capella-collab-manager contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import { TestBed } from '@angular/core/testing';
-
-import { SourceService } from './source.service';
-
-xdescribe('SourceService', () => {
-  let service: SourceService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(SourceService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
